+++ lib/keyboard_detection.dart
@@ -1,4 +1,3 @@
 library keyboard_detection;
 
 export 'src/keyboard_detection.dart';
-export 'src/keyboard_detection_controller.dart';